MM-220 fix: owner 권한 부여

+++ src/main/java/com/dev/moim/domain/moim/service/impl/MoimCommandServiceImpl.java
@@ -103,6 +103,10 @@ public class MoimCommandServiceImpl implements MoimCommandService {
 
         UserMoim userMoim = userMoimRepository.findByUserIdAndMoimId(user.getId(), moim.getId(), JoinStatus.COMPLETE).orElseThrow(() -> new MoimException(ErrorStatus.USER_NOT_MOIM_JOIN));
 
+        if (userMoim.getMoimRole().equals(MoimRole.OWNER)) {
+            throw new MoimException(ErrorStatus.OWNER_NOT_EXIT);
+        }
+
         userMoimRepository.delete(userMoim);
 
         Optional<UserMoim> ownerByMoim = userRepository.findOwnerByMoim(moim);
